# feat: optimize `RestMethodInfo`, reduce dictionary allocations and li…