--- a/Refit.Tests/RequestBuilder.cs
+++ b/Refit.Tests/RequestBuilder.cs
@@ -1062,8 +1062,8 @@ namespace Refit.Tests
             Assert.Equal("1", fixture.Headers["Api-Version"]);
 
             Assert.Equal(4, fixture.Headers.Count);
-            Assert.Equal(1, fixture.HeaderCollectionParameterMap.Count);
-            Assert.True(fixture.HeaderCollectionParameterMap.Contains(1));
+            Assert.True(fixture.HasHeaderCollection);
+            Assert.True(fixture.HeaderCollectionAt(1));
         }
 
         [Theory]
@@ -1085,8 +1085,8 @@ namespace Refit.Tests
             Assert.NotNull(fixture.BodyParameterInfo);
             Assert.Null(fixture.AuthorizeParameterInfo);
 
-            Assert.Equal(1, fixture.HeaderCollectionParameterMap.Count);
-            Assert.True(fixture.HeaderCollectionParameterMap.Contains(2));
+            Assert.True(fixture.HasHeaderCollection);
+            Assert.True(fixture.HeaderCollectionAt(2));
         }
 
         [Theory]
@@ -1110,8 +1110,8 @@ namespace Refit.Tests
             Assert.Null(fixture.BodyParameterInfo);
             Assert.Null(fixture.AuthorizeParameterInfo);
 
-            Assert.Equal(1, fixture.HeaderCollectionParameterMap.Count);
-            Assert.True(fixture.HeaderCollectionParameterMap.Contains(1));
+            Assert.True(fixture.HasHeaderCollection);
+            Assert.True(fixture.HeaderCollectionAt(1));
         }
 
         [Theory]
@@ -1141,8 +1141,8 @@ namespace Refit.Tests
             Assert.NotNull(fixture.BodyParameterInfo);
             Assert.Null(fixture.AuthorizeParameterInfo);
 
-            Assert.Equal(1, fixture.HeaderCollectionParameterMap.Count);
-            Assert.True(fixture.HeaderCollectionParameterMap.Contains(1));
+            Assert.True(fixture.HasHeaderCollection);
+            Assert.True(fixture.HeaderCollectionAt(1));
             Assert.Equal(2, fixture.BodyParameterInfo.Item3);
         }
 
@@ -1168,8 +1168,8 @@ namespace Refit.Tests
             Assert.Null(fixture.BodyParameterInfo);
 
             Assert.NotNull(fixture.AuthorizeParameterInfo);
-            Assert.Equal(1, fixture.HeaderCollectionParameterMap.Count);
-            Assert.True(fixture.HeaderCollectionParameterMap.Contains(2));
+            Assert.True(fixture.HasHeaderCollection);
+            Assert.True(fixture.HeaderCollectionAt(2));
         }
 
         [Theory]
@@ -1195,8 +1195,8 @@ namespace Refit.Tests
 
             Assert.Single(fixture.HeaderParameterMap);
             Assert.Equal("Authorization", fixture.HeaderParameterMap[1]);
-            Assert.Equal(1, fixture.HeaderCollectionParameterMap.Count);
-            Assert.True(fixture.HeaderCollectionParameterMap.Contains(2));
+            Assert.True(fixture.HasHeaderCollection);
+            Assert.True(fixture.HeaderCollectionAt(2));
 
             input = typeof(IRestMethodInfoTests);
             fixture = new RestMethodInfoInternal(
@@ -1220,8 +1220,8 @@ namespace Refit.Tests
 
             Assert.Single(fixture.HeaderParameterMap);
             Assert.Equal("Authorization", fixture.HeaderParameterMap[2]);
-            Assert.Equal(1, fixture.HeaderCollectionParameterMap.Count);
-            Assert.True(fixture.HeaderCollectionParameterMap.Contains(1));
+            Assert.True(fixture.HasHeaderCollection);
+            Assert.True(fixture.HeaderCollectionAt(1));
         }
 
         [Theory]
@@ -1253,8 +1253,8 @@ namespace Refit.Tests
 
             Assert.Single(fixture.HeaderParameterMap);
             Assert.Equal("X-PathMember", fixture.HeaderParameterMap[0]);
-            Assert.Equal(1, fixture.HeaderCollectionParameterMap.Count);
-            Assert.True(fixture.HeaderCollectionParameterMap.Contains(1));
+            Assert.True(fixture.HasHeaderCollection);
+            Assert.True(fixture.HeaderCollectionAt(1));
         }
 
         [Theory]
@@ -1274,8 +1274,8 @@ namespace Refit.Tests
             Assert.Null(fixture.BodyParameterInfo);
 
             Assert.Equal("baz", fixture.QueryParameterMap[2]);
-            Assert.Equal(1, fixture.HeaderCollectionParameterMap.Count);
-            Assert.True(fixture.HeaderCollectionParameterMap.Contains(1));
+            Assert.True(fixture.HasHeaderCollection);
+            Assert.True(fixture.HeaderCollectionAt(1));
         }
 
         [Theory]
@@ -1321,8 +1321,8 @@ namespace Refit.Tests
 
             Assert.Single(fixture.PropertyParameterMap);
 
-            Assert.Equal(1, fixture.HeaderCollectionParameterMap.Count);
-            Assert.True(fixture.HeaderCollectionParameterMap.Contains(0));
+            Assert.True(fixture.HasHeaderCollection);
+            Assert.True(fixture.HeaderCollectionAt(0));
         }
 
         [Theory]
@@ -1393,7 +1393,7 @@ namespace Refit.Tests
             Assert.Empty(fixture.HeaderParameterMap);
             Assert.NotNull(fixture.BodyParameterInfo);
             Assert.Null(fixture.AuthorizeParameterInfo);
-            Assert.Empty(fixture.HeaderCollectionParameterMap);
+            Assert.False(fixture.HasHeaderCollection);
 
             Assert.Equal("SomeProperty", fixture.PropertyParameterMap[2]);
         }
@@ -1420,7 +1420,7 @@ namespace Refit.Tests
             Assert.Empty(fixture.HeaderParameterMap);
             Assert.NotNull(fixture.BodyParameterInfo);
             Assert.Null(fixture.AuthorizeParameterInfo);
-            Assert.Empty(fixture.HeaderCollectionParameterMap);
+            Assert.False(fixture.HasHeaderCollection);
 
             Assert.Equal("SomeProperty", fixture.PropertyParameterMap[2]);
             Assert.Equal("SomeOtherProperty", fixture.PropertyParameterMap[3]);
@@ -1449,7 +1449,7 @@ namespace Refit.Tests
             Assert.Empty(fixture.HeaderParameterMap);
             Assert.Null(fixture.BodyParameterInfo);
             Assert.Null(fixture.AuthorizeParameterInfo);
-            Assert.Empty(fixture.HeaderCollectionParameterMap);
+            Assert.False(fixture.HasHeaderCollection);
 
             Assert.Equal("SomeProperty", fixture.PropertyParameterMap[1]);
         }
@@ -1477,7 +1477,7 @@ namespace Refit.Tests
             Assert.Empty(fixture.HeaderParameterMap);
             Assert.NotNull(fixture.BodyParameterInfo);
             Assert.Null(fixture.AuthorizeParameterInfo);
-            Assert.Empty(fixture.HeaderCollectionParameterMap);
+            Assert.False(fixture.HasHeaderCollection);
 
             Assert.Equal("SomeProperty", fixture.PropertyParameterMap[1]);
             Assert.Equal(2, fixture.BodyParameterInfo.Item3);