--- a/Refit/EnumerableExtensions.cs
+++ b/Refit/EnumerableExtensions.cs
@@ -19,6 +19,13 @@ internal static class EnumerableExtensions
     }
 }
 
+internal static class EmptyDictionary<TKey, TValue>
+{
+    private static Dictionary<TKey, TValue> Value = new Dictionary<TKey, TValue>();
+
+    internal static Dictionary<TKey, TValue> Get() => Value;
+}
+
 internal enum EnumerablePeek
 {
     Empty,