--- a/Refit/RequestBuilderImplementation.cs
+++ b/Refit/RequestBuilderImplementation.cs
@@ -689,8 +689,8 @@ namespace Refit
                                                 Uri.EscapeDataString(
                                                     settings.UrlParameterFormatter.Format(
                                                         s,
-                                                        restMethod.ParameterInfoMap[i],
-                                                        restMethod.ParameterInfoMap[i].ParameterType
+                                                        restMethod.ParameterInfoArray[i],
+                                                        restMethod.ParameterInfoArray[i].ParameterType
                                                     ) ?? string.Empty
                                                 )
                                         )
@@ -702,8 +702,8 @@ namespace Refit
                                 replacement = Uri.EscapeDataString(
                                     settings.UrlParameterFormatter.Format(
                                         param,
-                                        restMethod.ParameterInfoMap[i],
-                                        restMethod.ParameterInfoMap[i].ParameterType
+                                        restMethod.ParameterInfoArray[i],
+                                        restMethod.ParameterInfoArray[i].ParameterType
                                     ) ?? string.Empty
                                 );
                             }
@@ -738,7 +738,7 @@ namespace Refit
                     }
 
                     //if header collection, add to request headers
-                    if (restMethod.HeaderCollectionParameterMap.Contains(i))
+                    if (restMethod.HeaderCollectionAt(i))
                     {
                         if (param is IDictionary<string, string> headerCollection)
                         {
@@ -776,7 +776,7 @@ namespace Refit
                     // or if is an object bound to the path add any non-path bound properties to query string
                     // or if it's an object with a query attribute
                     var queryAttribute = restMethod
-                        .ParameterInfoMap[i]
+                        .ParameterInfoArray[i]
                         .GetCustomAttribute<QueryAttribute>();
                     if (
                         !restMethod.IsMultipart
@@ -897,7 +897,7 @@ namespace Refit
                 queryParamsToAdd.AddRange(
                     ParseQueryParameter(
                         param,
-                        restMethod.ParameterInfoMap[i],
+                        restMethod.ParameterInfoArray[i],
                         restMethod.QueryParameterMap[i],
                         attr
                     )
@@ -913,7 +913,7 @@ namespace Refit
                     queryParamsToAdd.AddRange(
                         ParseQueryParameter(
                             kvp.Value,
-                            restMethod.ParameterInfoMap[i],
+                            restMethod.ParameterInfoArray[i],
                             path,
                             attr
                         )