@ChrisPulman 2024-06-29
Now I see what happened before, as you click merge the browser then scrolls after a short delay making you think your clicking confirm, but the scrolling pushes the Close button into the same place 😖